fix: 병합 충돌 해결 및 swagger DTO 명명 영어로 변경
