+++ src/main/java/com/ceos/bankids/dto/AchievedChallengeListDTO.java
@@ -1,6 +1,5 @@
 package com.ceos.bankids.dto;
 
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import java.util.List;
 import lombok.EqualsAndHashCode;
@@ -8,7 +7,6 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
-@ApiModel(value = "완주한 돈길 리스트 DTO")
 @Getter
 @Setter
 @EqualsAndHashCode
