+++ src/main/java/com/ceos/bankids/dto/ChallengeDTO.java
@@ -4,7 +4,6 @@ import com.ceos.bankids.constant.ChallengeStatus;
 import com.ceos.bankids.domain.Challenge;
 import com.ceos.bankids.domain.Comment;
 import com.fasterxml.jackson.annotation.JsonFormat;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import java.sql.Timestamp;
 import java.util.List;
@@ -12,7 +11,6 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
-@ApiModel(value = "돈길 관련 API response DTO")
 @Getter
 @ToString
 @EqualsAndHashCode
