+++ src/main/java/com/ceos/bankids/dto/KidAchievedChallengeListDTO.java
@@ -1,13 +1,11 @@
 package com.ceos.bankids.dto;
 
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
-@ApiModel(value = "자녀의 완주한 돈길 리스트 DTO")
 @Getter
 @Setter
 @EqualsAndHashCode
