+++ src/main/java/com/ceos/bankids/dto/KidWeekDTO.java
@@ -1,13 +1,11 @@
 package com.ceos.bankids.dto;
 
 import com.ceos.bankids.domain.Kid;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
-@ApiModel(value = "자녀의 주차 정보")
 @Getter
 @ToString
 @EqualsAndHashCode
