+++ src/main/java/com/ceos/bankids/dto/NotificationListDTO.java
@@ -1,13 +1,11 @@
 package com.ceos.bankids.dto;
 
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import java.util.List;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
-@ApiModel(value = "알림 리스트로 줄 때 DTO")
 @Getter
 @ToString
 @EqualsAndHashCode
