+++ src/main/java/com/ceos/bankids/dto/PreSignedDTO.java
@@ -1,6 +1,5 @@
 package com.ceos.bankids.dto;
 
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import java.net.URL;
 import lombok.EqualsAndHashCode;
@@ -8,7 +7,6 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
-@ApiModel(value = "preSignedUrl 받아오는 DTO")
 @Setter
 @Getter
 @ToString
