+++ src/main/java/com/ceos/bankids/dto/WeekDTO.java
@@ -1,12 +1,10 @@
 package com.ceos.bankids.dto;
 
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
-@ApiModel(value = "주차 정보 DTO")
 @Getter
 @ToString
 @EqualsAndHashCode
